Fixed the problem with forecast and unit conversion
Before when clicking on Fahrenheit link and then searching for another city or clicking current location button the temperature was updating but the fahrenheit link was marked as active. I removed the active class from fahrenheit link and added it to celsius link since celsius is the default temperature here. I also fixed the issue with forecast. Before when one searched for a city, the app displayed correct weather and foecast for the given city, but then when clicking on the current location button the upcoming hours forecast wasn't updating (while the current weather info was correct). I updated the api url and called the forecast function from inside of the showPositon(position) function.